# uint256: optimize AddMod