@holiman 2024-05-10
Can confirm: 
```
[user@work uint256]$ benchstat addmod.1 addmod.2
name                     old time/op  new time/op  delta
AddMod/small/uint256-8   21.0ns ±44%  17.0ns ±43%     ~     (p=0.052 n=10+10)
AddMod/mod64/uint256-8   28.6ns ±36%  26.6ns ±44%     ~     (p=0.481 n=10+10)
AddMod/mod128/uint256-8  56.1ns ±43%  53.7ns ±49%     ~     (p=0.810 n=10+10)
AddMod/mod192/uint256-8  62.0ns ±44%  62.2ns ±44%     ~     (p=0.971 n=10+10)
AddMod/mod256/uint256-8  25.4ns ±25%  12.0ns ±44%  -52.74%  (p=0.000 n=10+9)
```
Neat!